requirements.txt
streamlit==1.34.0
google-generativeai==0.3.1
streamlit-turnstile==0.0.4
